% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim-chpi-serob.R
\name{cbpcode}
\alias{cbpcode}
\title{CBP model fit}
\usage{
cbpcode(y, x, probud = 0.5, alpha = 1, beta = 1)
}
\arguments{
\item{y}{a data frame or tibble}

\item{x}{asdasd}

\item{probud}{probability of group 1}

\item{alpha}{mu1}

\item{beta}{effect}
}
\value{
Returns a list with the following elements:.
}
\description{
Fit a mixture model using moments. Only one component prototype
}
\examples{

p <- 0.7  #
N <- 100
x <- rnorm(N)
x2 <- rnorm(N)
y <- rnorm(N, mean=x)
y[1:(N*p)] <- rnorm(N*p, mean=0, sd=.25)

cbpcode(y, x)

}
\author{
Christian Pipper \email{pipper@sund.ku.dk}
}
\keyword{manip}
