% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{gaussian_mixture_model}
\alias{gaussian_mixture_model}
\title{Fit 2-component Gaussian mixture model}
\usage{
gaussian_mixture_model(data, formula, B = 200, ...)
}
\arguments{
\item{data}{a data frame or tibble}

\item{formula}{a formula model formula for the relationship}

\item{B}{a positive integer giving the number of bootstrap samples used to obtain the estimates of the variance associated with the parameter estimates.}

\item{...}{Additional arguments passed on to the fitting process.}
}
\value{
Returns a list with the following elements: ...
}
\description{
Fit a 2-component Gaussian mixture model.
}
\examples{

p <- 0.7  #
N <- 100
x <- rnorm(N)
x2 <- rnorm(N)
y <- rnorm(N, mean=x)
y[1:(N*p)] <- rnorm(N*p, mean=0, sd=.25)

DF <- data.frame(x, x2, y)

gaussian_mixture_model(DF, y ~ x + x2)

}
\author{
Claus Ekstrom \email{ekstrom@sund.ku.dk} and Christian Pipper \email{pipper@sund.ku.dk}
}
\keyword{manip}
