% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmm.R
\name{moment_mixture_model}
\alias{moment_mixture_model}
\title{Mixture modeling using moments}
\usage{
moment_mixture_model(data, formula, weights = c("equal", "square"),
  maxiter = 5)
}
\arguments{
\item{data}{a data frame or tibble}

\item{formula}{a formula model formula for the relationship}

\item{weights}{a string that defines the weights used for the estimation using weighted regression. Possibilities are "equal" (the default) for unweighted regression estimation and "square" for squared linear predictor estimation.}

\item{maxiter}{integer}
}
\value{
Returns a list with the following elements:

* coefficients The estimates of the parameters for 
* alpha
* mu
* covariance
* call
}
\description{
Fit a mixture model using moments
}
\examples{

p <- 0.7  #
N <- 100
x <- rnorm(N)
x2 <- rnorm(N)
y <- rnorm(N, mean=x)
y[1:(N*p)] <- rnorm(N*p, mean=0, sd=.25)

DF <- data.frame(x, x2, y)

moment_mixture_model(DF, y ~ x + x2)

}
\author{
Claus Ekstrom \email{ekstrom@sund.ku.dk} and Christian Pipper \email{pipper@sund.ku.dk}
}
\keyword{manip}
