% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mommix-package.R
\docType{package}
\name{mommix}
\alias{mommix}
\alias{mommix-package}
\title{Functions for mixture estimation based on moments}
\description{
Functions for mixture estimation based on moments.
}
\details{
\tabular{ll}{ Package: \tab mommix\cr Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2012-03-29\cr License: \tab GPL-2\cr } % ~~ An overview of
how to use the package, including the most important ~~ % ~~ functions ~~
}
\references{
Some paper
}
\author{
Claus Thorn Ekstrom \email{ekstrom@sund.ku.dk}\cr Maintainer: Claus Thorn Ekstrom
\email{ekstrom@sund.ku.dk}
}
\keyword{package}
