% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mgrwc}
\alias{mgrwc}
\title{Gaussian mixture of regression models with contamination}
\usage{
mgrwc(y, X, maxit = 400L, tol = 1e-07)
}
\arguments{
\item{y}{A vector of outcomes.}

\item{X}{A design matrix of regressor variables. Must have the same number of rows as the length of y.}

\item{maxit}{the maximum number of iterations to use for the EM algorithm}

\item{tol}{the tolerance used for determining convergence}
}
\value{
A list with the variables: coefficients, stderr, and tstat that gives the slope estimate, the corresponding standard error, and their ratio for each column in x.
}
\description{
Fast computation of simple regression slopes for each predictor represented by a column in a matrix
}
\details{
Missing values (NA, Inf, NaN) are completely disregarded and pairwise complete cases are used for the analysis.
}
\examples{
x <- rnorm(1000)
y <- rnorm(1000, mean=c(x[1:700], rep(0, 300)), sd=rep(c(1,2), times=c(700,300)))
mgrwc(y, cbind(rep(1, 1000), x))

}
\author{
Claus Ekstrom <ekstrom@sund.ku.dk>
}
